scripts/requirements.txt
# DocuSearch_AI Python Dependencies

# Image processing and EXIF extraction
Pillow>=10.0.0
piexif>=1.1.3

# HTTP requests for API calls
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Retry logic for API calls
tenacity>=8.2.0
